+++ src/main/java/utils/CrewAttendanceFileReader.java
@@ -1,4 +1,4 @@
-package domain;
+package utils;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
